+++ src/App.tsx
@@ -18,6 +18,7 @@ import { StripeView } from "./components/views/StripeView";
 import { AiInsightsView } from "./components/views/AiInsightsView";
 import { EmailMarketingView } from "./components/views/EmailMarketingView";
 import { ProductsView } from "./components/views/ProductsView";
+import { CeoNotesView } from "./components/views/CeoNotesView";
 import { InboxView } from "./components/views/InboxView";
 import { ReportsView } from "./components/views/ReportsView";
 import { SettingsView } from "./components/views/SettingsView";
@@ -115,6 +116,8 @@ const CRMMainContent: React.FC = () => {
         return <EmailMarketingView />;
       case "products":
         return <ProductsView />;
+      case "ceonotes":
+        return <CeoNotesView />;
       case "inbox":
         return <InboxView />;
       case "reports":